fix SafeAreaView and directory name
