+++ app/(tabs)/comment_challenge.tsx
@@ -1,10 +1,11 @@
-import { StyleSheet, Text, SafeAreaView, View, TextInput, ScrollView } from "react-native";
+import { StyleSheet, Text, View, TextInput, ScrollView, } from "react-native";
 import { useState } from "react";
 import CalenderView from "@/components/CalendarView";
 import SmallButton from "@/components/SmallButton";
 import SectionButton from "@/components/SectionButton";
 import colors from "@/constants/Colors";
 import spacing from "@/constants/spacing";
+import { SafeAreaView } from "react-native-safe-area-context";
 
 export default function CommentScreen() {
   const [selected, setSelected] = useState(new Date(Date.now()) + "");
@@ -16,7 +17,7 @@ export default function CommentScreen() {
 
   return (
     <>
-      <SafeAreaView style={styles.container}>
+        <SafeAreaView style={styles.container} edges={{bottom: "off", top:"additive"}}>
         <SmallButton href="./chat" text="Trainer" />
         <ScrollView>
           <CalenderView
