+++ app/(tabs)/schedule.tsx
@@ -1,127 +1,186 @@
-import React, { FC, useState, useEffect } from "react";
-// import TextField from "@/components/TextField";
 import {
-  TextInput,
-  StyleSheet,
-  Text,
-  View,
-  Pressable,
-  ScrollView,
-} from "react-native";
-import colors from "@/constants/Colors";
-import Fullline from "@/components/Fullline";
-import { SafeAreaView } from "react-native-safe-area-context";
-
-export default function ScheduleScreen() {
-  const [title, setTitle] = useState("");
-  const [date, setDate] = useState(new Date(Date.now()));
-  const [color, setColor] = useState("FFFFFF99");
-  const [content, setContent] = useState("");
-
-  const months = [
-    "1월",
-    "2월",
-    "3월",
-    "4월",
-    "5월",
-    "6월",
-    "7월",
-    "8월",
-    "9월",
-    "10월",
-    "11월",
-    "12월",
-  ];
-  const days = ["일", "월", "화", "수", "목", "금", "토"];
-
-  const formattedDate = `${date.getFullYear()}년 ${months[date.getMonth()]} ${date.getDate()}일 (${days[date.getDay()]})`;
-
-  useEffect(() => {
-    setDate(date);
-    setTitle(title);
-    setColor(color);
-    setContent(content);
-    return () => {
-      setDate(date);
-      setTitle(title);
-      setColor(color);
-      setContent(content);
+    View,
+    ScrollView,
+    StyleSheet,
+    Text,
+    TextInput,
+    TouchableOpacity,
+    Image,
+  } from "react-native";
+  import { useState } from "react";
+  import { useAuth0 } from "react-native-auth0";
+  
+  import ScrollButton from "@/components/ScrollButton";
+  import SectionButton from "@/components/SectionButton";
+  import SmallButton from "@/components/SmallButton";
+  import SwitchView from "@/components/SwitchView";
+  import colors from "@/constants/Colors";
+  import spacing from "@/constants/spacing";
+  import CalendarView from "@/components/CalendarView";
+  import { SwipeListView } from "react-native-swipe-list-view";
+  import { SafeAreaView } from "react-native-safe-area-context";
+  
+  interface ScheduleItem {
+    id: string;
+    title: string;
+    contents: string;
+    completed: boolean;
+  }
+  
+  export default function MainScreen() {
+    /* hacky way to convert to string */
+    const modifyicon = require("@/assets/images/editicon.png");
+    const deleteicon = require("@/assets/images/trashicon.png");
+    const circle = require("@/assets/images/whitecircle.png");
+    const [selected, setSelected] = useState(new Date(Date.now()) + "");
+  
+    const [data, setData] = useState<ScheduleItem[]>([
+      { id: "1", title: "Item 1", contents: "content1", completed: false },
+      { id: "2", title: "Item 2", contents: "content2", completed: false },
+      { id: "3", title: "Item 3", contents: "content3", completed: false },
+      // Add more items as needed
+    ]);
+  
+    const { clearSession } = useAuth0();
+  
+    const onLogout = async () => {
+      try {
+        await clearSession();
+      } catch (e) {
+        console.log("Log out cancelled");
+      }
     };
-  });
-  return (
-    <>
-      <SafeAreaView style={styles.container}>
-        <ScrollView>
-          <TextInput
-            value={title}
-            onChangeText={(value: string) => setTitle(value)}
-            placeholder="할 일 제목"
-            style={styles.title}
-            textAlignVertical="center"
-            placeholderTextColor={colors.main.background}
-            maxLength={30}
-          />
-          <View style={styles.button}>
-            <Pressable>
-              <Text style={styles.buttontext}>컬러</Text>
-            </Pressable>
-          </View>
-          <Text style={styles.datestyle}>
-            {` ${date.getMonth() + 1}월 ${date.getDate()}일 ${days[date.getDay()]}요일`}
+  
+    const renderItem = ({ item }: { item: ScheduleItem }) => (
+      <View style={styles.row}>
+        <View style={styles.rowcontent}>
+          <Text
+            style={{
+              color: colors.white.background,
+              fontSize: 18,
+              fontWeight: "normal",
+            }}
+          >
+            {item.title}
           </Text>
-          {/* ${date.getFullYear()}년 */}
-          <Fullline />
-          <TextInput
-            value={content}
-            onChangeText={(value: string) => setContent(value)}
-            placeholder="내용을 입력하세요."
-            style={styles.content}
-            multiline={true}
-            placeholderTextColor={"black"}
+          <View style={styles.interval}>
+            <Image source={circle} style={{ height: 8, width: 8 }} />
+            <Text
+              style={{
+                color: colors.white.background,
+                fontSize: 15,
+                fontWeight: "normal",
+              }}
+            >
+              {item.contents}
+            </Text>
+          </View>
+        </View>
+        <SwitchView />
+        <SwipeListView />
+      </View>
+    );
+  
+    const renderHiddenItem = () => (
+      <View style={styles.rowBack}>
+        <TouchableOpacity onPress={() => console.log("left button click")}>
+          <Image source={modifyicon} style={{ height: 20, width: 20 }} />
+        </TouchableOpacity>
+        <TouchableOpacity onPress={() => console.log("right button click")}>
+          <Image source={deleteicon} style={{ height: 20, width: 20 }} />
+        </TouchableOpacity>
+      </View>
+    );
+  
+    return (
+      <SafeAreaView style={styles.container} edges={{bottom: "off", top: "additive"}}>
+        <ScrollView>
+          <SmallButton href="./chat" text="Trainer" />
+          <CalendarView
+            selectday={selected}
+            onDayPress={(day: { dateString: string }) =>
+              setSelected(day.dateString)
+            }
           />
+          <View style={styles.content}>
+            <Text style={styles.selectday}>{formatDate(new Date(selected))}</Text>
+            <View style={styles.section}>
+              <SectionButton href="../addSchedule" text="Schedule +" />
+            </View>
+              <View style={styles.schedulelist}>
+                <SwipeListView
+                  data={data}
+                  renderItem={renderItem}
+                  renderHiddenItem={renderHiddenItem}
+                  leftOpenValue={50}
+                  rightOpenValue={-50}
+                />
+              </View>
+          </View>
         </ScrollView>
       </SafeAreaView>
-    </>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.white.background,
-  },
-  title: {
-    fontSize: 33,
-    fontWeight: "bold",
-    color: colors.main.background,
-    marginLeft: "10%",
-    marginRight: "10%",
-  },
-  button: {
-    width: 65,
-    height: 20,
-    backgroundColor: colors.main.background,
-    borderRadius: 15,
-    justifyContent: "center",
-    alignItems: "center",
-    marginLeft: "10%",
-    marginTop: "5%",
-  },
-  buttontext: {
-    color: colors.white.background,
-    fontWeight: "bold",
-    fontSize: 10,
-  },
-  datestyle: {
-    fontSize: 15,
-    fontWeight: "bold",
-    color: colors.main.background,
-    marginTop: "5%",
-    marginLeft: "10%",
-  },
-  content: {
-    fontSize: 15,
-    marginLeft: "10%",
-    marginRight: "10%",
-  },
-});
+    );
+  }
+  
+  function formatDate(d: Date) {
+    const dayNames = ["Sun", "Mon", "Tue", "Wed", "Thur", "Fri", "Sat"];
+    return dayNames[d.getDay()] + " " + d.getDate();
+  }
+  
+  const styles = StyleSheet.create({
+    container:{
+      flex: 1,
+      backgroundColor: "#F7F9FC",
+    },
+    content: {
+      // flex: 2,
+      marginTop: "5%",
+    },
+    interval: {
+      flexDirection: "row",
+      alignItems: "center",
+      marginTop: "7%",
+      marginLeft: "5%",
+      gap: spacing.s,
+    },
+    selectday: {
+      marginLeft: "5%",
+      fontSize: 25,
+      fontWeight: "bold",
+      color: colors.main.background,
+    },
+    section: {
+      flex: 1,
+    },
+    schedulelist: {
+      alignItems: "center",
+    },
+    row: {
+      flexDirection: "row",
+      justifyContent: "space-around",
+      alignItems: "center",
+      textAlign: "center",
+      height: 80,
+      width: 320,
+      marginTop: "5%",
+      borderRadius: spacing.m,
+      backgroundColor: colors.sub1.background,
+    },
+    rowcontent: {
+        flexDirection: "column",
+      },
+    rowBack: {
+      alignItems: "center",
+      backgroundColor: colors.white.background,
+      flex: 1,
+      height: 80,
+      width: 320,
+      borderRadius: spacing.s,
+      marginTop: "5%",
+      flexDirection: "row",
+      justifyContent: "space-between",
+      paddingLeft: spacing.m,
+      paddingRight: spacing.m,
+    },
+  });
+
